Rename BottomInfoSection
