+++ front/app/containers/Admin/pages-menu/containers/HeroBanner/HeaderImageDropzone.tsx
